[ralph-7] S-4: Fix _fidelity_to_task_idx dtype to torch.long
Task indices for IndexKernel must be integer type. Changed dtype from
fidelity_values.dtype (float64) to torch.long. Added dtype assertions
to both existing tests.

+++ gopro/04_gpbo_loop.py
@@ -391,7 +391,7 @@ def _fidelity_to_task_idx(
     # exact float constants (0.0, 0.5, 1.0) assigned by merge_multi_fidelity_data.
     task_idx = torch.tensor(
         [fidelity_map[f] for f in fidelity_values.tolist()],
-        dtype=fidelity_values.dtype,
+        dtype=torch.long,
     )
     return task_idx, fidelity_map
